Rename from fargate to ecs
