+++ readthedocs/projects/apps.py
@@ -15,7 +15,7 @@ class ProjectsConfig(AppConfig):
         import readthedocs.projects.notifications  # noqa
         import readthedocs.projects.signals  # noqa
         import readthedocs.projects.tasks.builds  # noqa
-        import readthedocs.projects.tasks.fargate  # noqa
+        import readthedocs.projects.tasks.ecs  # noqa
         import readthedocs.projects.tasks.search  # noqa
         import readthedocs.projects.tasks.utils  # noqa
